add heatmap to main page
