+++ src/components/NavHeader.js
@@ -1,6 +1,6 @@
 const NavHeader = () => {
     return (
-        <div className='bg-slate-400 block h-14'>Nav Header</div>
+        <div>Nav Header</div>
     )
 }
